+++ src/main/java/frc/robot/TeleopCommander.java
@@ -6,7 +6,7 @@ import edu.wpi.first.math.trajectory.Trajectory.State;
 import edu.wpi.first.wpilibj.XboxController;
 //import frc.robot.Subsystems.Arm.armDesiredPos;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import frc.robot.trajectory.RotationSequence;
+import frc.robot.utils.trajectory.RotationSequence;
 
 public class TeleopCommander implements RobotCommander {
     private static XboxController driver;
