+++ src/main/java/frc/robot/utils/Interpolation/Interpolable.java
@@ -1,4 +1,4 @@
-package frc.robot.utils;
+package frc.robot.utils.Interpolation;
 
 /**
  * Interpolable is an interface used by an Interpolating Tree as the Value type.
