+++ src/main/java/frc/robot/utils/trajectory/CustomTrajectoryGenerator.java
@@ -5,7 +5,7 @@
 // license that can be found in the LICENSE file at
 // the root directory of this project.
 
-package frc.robot.trajectory;
+package frc.robot.utils.trajectory;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
