+++ src/main/java/frc/robot/utils/trajectory/Waypoint.java
@@ -5,7 +5,7 @@
 // license that can be found in the LICENSE file at
 // the root directory of this project.
 
-package frc.robot.trajectory;
+package frc.robot.utils.trajectory;
 
 import static edu.wpi.first.util.ErrorMessages.requireNonNullParam;
